@@ -5,9 +5,12 @@ const passport = require('passport');
 const userService = require('../services/user.service');
 const secret = process.env.SECRET || 'tasmanianDevil';
 const isAdmin = require('../middlewares/is-admin.middleware').isAdmin;
-var api_key = 'key-c0515d395d6d5188df64af9c2f783194';
-var DOMAIN = 'sandbox171d4d23ca894d458d322157f279255a.mailgun.org';
-var mailgun = require('mailgun-js')({apiKey: api_key, domain: DOMAIN});
+const config = require('../config/config.js');
+const api_key = config.get('mailgun_api_key');
+const DOMAIN = config.get('mailgun_domain');
+const mailgun = require('mailgun-js')({apiKey: api_key, domain: DOMAIN});
+const mailTemplate = require('../templates/mail');
+var Chance = require('chance');
 // Models
 const User = require('../models/user.model').users;
 
@@ -43,13 +46,11 @@ router.post('/', (req, res, next) => {
   const user = {name,email,phone,github,password};
   userService.save(user)
     .then((user) => {
-      var data = {
-        from: 'Angular Pakistan <no-reply@ngpakistan.com>',
-        to: user.email,
-        subject: 'Verify your account',
-        text: `Dear ${user.name},\nClick on the following link to verify your account: http://ngpakistan.com/#/verify/${user._id}`
-      };
-      mailgun.messages().send(data, function (error, body) {
+      const chance = new Chance();
+      const token = chance.word({length: 16});
+      user.token = token;
+      const data = mailTemplate.createVerificationMail(user.email, user.name, token);
+      mailgun.messages().send(data, (error, body) => {
           if(error){
               console.log(error);
               res.json({success: false, error: error});
@@ -58,6 +59,7 @@ router.post('/', (req, res, next) => {
           console.log('Successfully sent!');
           res.json({success: true});
       });
+      user.save();
     })
     .catch(err => {
       if(err)
@@ -75,11 +77,11 @@ router.post('/login', (req, res, next) => {
       } else {
         if (user.validPassword(password)) {
           if (user.verified) {
-            var payload = {id: user.id};
+            var payload = {id: user.id, name: user.name};
             if (user.admin) {
               payload.admin = true;
             }
-            var token = jwt.sign(payload, secret, { expiresIn: '24h'});
+            var token = jwt.sign(payload, secret, { expiresIn: '7d'});
             res.json({success: true, token: 'bearer ' + token});
           } else {
             res.json({error: `Your email hasn't been verified yet.
@@ -97,17 +99,22 @@ router.post('/logout', passport.authenticate('jwt', { session: false }), (req, r
   res.sendStatus(200);
 });
 
-router.get('/verify/:id', (req, res) => {
-  const id = req.params.id;
+router.get('/verify', (req, res) => {
+  const email = req.query.email;
+  const token = req.query.token;
 
-  userService.getUser(id).then(user => {
+  userService.getUserByEmail(email).then(user => {
     if(user){
       if(user.verified){
         res.json({success: false});
       } else {
-        user.verified = true;
-        user.save();
-        res.json({success: true});
+        if(user.token === token) {
+          user.verified = true;
+          user.save();
+          res.json({success: true});
+        } else {
+          res.json({success: false});
+        }
       }
     } else {
       res.sendStatus(404)
