@@ -2,8 +2,6 @@ import { PressKitComponent } from './presskit.component';
 
 export const routes = [
   {
-    path: '', children: [
-    { path: '', component: PressKitComponent }
-  ]
+    path: 'presskit',  component: PressKitComponent
   }
 ];
